+++ PKHeX.Core/PersonalInfo/Table/PersonalTable7GG.cs
@@ -9,8 +9,8 @@ public sealed class PersonalTable7GG : IPersonalTable, IPersonalTable<PersonalIn
 {
     private readonly PersonalInfo7GG[] Table;
     private const int SIZE = PersonalInfo7GG.SIZE;
-    private const int MaxSpecies = Legal.MaxSpeciesID_7b;
-    public int MaxSpeciesID => MaxSpecies;
+    private const ushort MaxSpecies = Legal.MaxSpeciesID_7b;
+    public ushort MaxSpeciesID => MaxSpecies;
 
     public PersonalTable7GG(ReadOnlySpan<byte> data)
     {
