+++ PKHeX.Core/PersonalInfo/Table/PersonalTable8BDSP.cs
@@ -9,8 +9,8 @@ public sealed class PersonalTable8BDSP : IPersonalTable, IPersonalTable<Personal
 {
     private readonly PersonalInfo8BDSP[] Table;
     private const int SIZE = PersonalInfo8BDSP.SIZE;
-    private const int MaxSpecies = Legal.MaxSpeciesID_8b;
-    public int MaxSpeciesID => MaxSpecies;
+    private const ushort MaxSpecies = Legal.MaxSpeciesID_8b;
+    public ushort MaxSpeciesID => MaxSpecies;
 
     public PersonalTable8BDSP(ReadOnlySpan<byte> data)
     {
